requirements.txt
Flask==2.1.0
psycopg2-binary==2.9.3
yfinance==0.1.63
scikit-learn==1.0.2
pandas==1.4.2
matplotlib==3.5.1
